implemented purchasing multiple items of same type

+++ client/src/components/ecommerce/CheckoutProduct.js
@@ -3,27 +3,50 @@ import '../../css/CheckoutProduct.css';
 import db from '../../firebase';
 import {useAuth} from '../../contexts/AuthContext';
 
-function CheckoutProduct({productId, title, price, image, rating, setLength, setTotal}) {
+function CheckoutProduct({productId, title, price, image, rating, quantity, setLength, setTotal}) {
     const {currentUser} = useAuth();
 
     const removeFromBasket = () => {
-        db.collection("users").doc(currentUser.uid).collection("basketItems").doc(productId).delete().then(() => {
-            console.log("Item successfully deleted!");
-            db.collection("users").doc(currentUser.uid).get().then(docc => {
-                const data = docc.data()
-                db.collection("users").doc(currentUser.uid).update({
-                    subtotal: data.subtotal - price,
-                    noItems: data.noItems - 1
+        if(quantity > 1) {
+            db.collection("users").doc(currentUser.uid).collection("basketItems").doc(productId).update({
+                itemQuantity: quantity - 1
+            }).then(() => {
+                console.log("Item successfully deleted!");
+                db.collection("users").doc(currentUser.uid).get().then(docc => {
+                    const data = docc.data()
+                    db.collection("users").doc(currentUser.uid).update({
+                        subtotal: data.subtotal - price,
+                        noItems: data.noItems - 1
+                    })
+                    setLength(data.noItems - 1);
+                    setTotal(data.subtotal - price);
+                    if (data.noItems - 1 === 0) {
+                        setTotal(0);
+                    }
                 })
-                setLength(data.noItems - 1);
-                setTotal(data.subtotal - price);
-                if (data.noItems - 1 === 0) {
-                    setTotal(0);
-                }
-            })
-        }).catch((error) => {
-            console.error("Error removing item: ", error);
-        });
+            }).catch((error) => {
+                console.error("Error removing item: ", error);
+            });
+        }
+        else {
+            db.collection("users").doc(currentUser.uid).collection("basketItems").doc(productId).delete().then(() => {
+                console.log("Item successfully deleted!");
+                db.collection("users").doc(currentUser.uid).get().then(docc => {
+                    const data = docc.data()
+                    db.collection("users").doc(currentUser.uid).update({
+                        subtotal: data.subtotal - price,
+                        noItems: data.noItems - 1
+                    })
+                    setLength(data.noItems - 1);
+                    setTotal(data.subtotal - price);
+                    if (data.noItems - 1 === 0) {
+                        setTotal(0);
+                    }
+                })
+            }).catch((error) => {
+                console.error("Error removing item: ", error);
+            });
+        }
     };
     return (
         <div className="checkoutProduct">
@@ -36,6 +59,8 @@ function CheckoutProduct({productId, title, price, image, rating, setLength, set
                     <small>$</small>
                     <strong>{price}</strong>
                 </p>
+                <p>Quantity: &nbsp;
+                <strong>{quantity}</strong></p>
                 <div className="checkoutProduct_rating">
                     {Array(rating)
                         .fill()
